Refactor: 캐릭터 선택 모달 로직 전부 제거, 프로필섹션 훅으로 넘김
