+++ src/components/pages/Mypage/components/CharSelectModal.jsx
@@ -1,31 +1,9 @@
-import { useState } from 'react';
-import { useRecoilState } from 'recoil';
-
-import { UserService } from 'network';
-import { userState } from 'store/user';
 import { ModalPortal, profileUtils } from '../utils';
 import MypageButton from 'components/pages/Mypage/components/MypageButton';
 
 import { StyledCharSelectModal } from '../styles';
 
-const CharSelectModal = ({ userInfo, setUserInfo, setIsModalOpen }) => {
-  const PICTURE_DIR = '/assets/CharacterWhiteBG/';
-  const charID = userInfo.character ?? 0;
-
-  const handleCharClick = async id => {
-    try {
-      const response = await UserService.updateUser({ character: id });
-      window.alert('캐릭터 변경 완료');
-      setUserInfo({ ...userInfo, character: id });
-    } catch (_) {
-      window.alert('캐릭터 변경 실패, 관리자에게 문의하세요');
-    }
-  };
-
-  const handleCloseClick = () => {
-    setIsModalOpen(false);
-  };
-
+const CharSelectModal = ({ userInfo, handleCharClick, handleCloseClick }) => {
   return (
     <ModalPortal>
       <StyledCharSelectModal>
@@ -38,9 +16,9 @@ const CharSelectModal = ({ userInfo, setUserInfo, setIsModalOpen }) => {
                 <li key={char.id}>
                   <button onClick={() => handleCharClick(char.id)}>
                     <img
-                      className={char.id === charID ? 'selected' : ''}
+                      className={char.id === userInfo.character ? 'selected' : ''}
                       alt="profile"
-                      src={PICTURE_DIR + char.image}
+                      src={'/assets/CharacterWhiteBG/' + char.image}
                     />
                   </button>
                 </li>
